feat: build directory setting & reuse (#181)
Support for custom build directory, opt-in. Using `FORCE` for the
moment. Doesn't try to rewrite the paths for moving CMake yet, but the
info for this is there in the new json file.

Maybe this shouldn't have a `cmake.` prefix? Edit: yes, looks like we
could sync-up with meson-python if we just use `build-dir`, so let's go
with that. See https://github.com/mesonbuild/meson-python/issues/275 for
`builddir` -> `build-dir` there.

- [x] TODO: this should be able to (maybe via substitution?) have a
unique path per interpreter.

Might adjust this later. My thought is we can make this opt-in for now,
then if we make it opt-out, we can protect the change on a minimum
version selection.

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>