feat: support dynamic metadata
This is a first pass at a scheme for supporting dynamic metadata in a
generic way. It keeps everything as simple as possible by using entry-points
with a very simple signature to provide each value.

Scikit-build-core's settings have been extended to have an additional metadata
field which is just an optional list mapping string keys to string values.
For each entry in this list, if the key is also included in the
project.dynamic list, scikit-build-core will look up an entry-point from the
"skbuild" group where the name matches the value. This entry-point should be a
function that accepts the path to pyproject.toml (this seemed like a minimum
but the final chosen signature is very much up for debate) and returns the
entry that should be inserted into the pyproject["project"][key] entry in the
loaded toml config.

Also included are a couple of core "plugins" which provide entry-points for
setuptools_scm and hatch-fancy-pypi-readme, essentially just adapters as
neither tool provides an explicit function matching the proposed entry-point
signature.

For tests, I have created a fixture which mocks the entrypoints list with fake
functions to give more direct control over what values we want to test. I have
duplicated most of the "simplest_c" test package to allow the complete build
process to be applied, but it is also possible to test with only a
pyproject.toml file, and in fact most of the tests I have currently implemented
do only use the pyproject.toml file and stop at testing the calculated
metadata, although this can of course be changed. There is also a simple test
for the setuptools_scm and hatch-fancy-pypi-readme plugins to make sure they
are working correctly.

Adding setuptools_scm as a test dependency to test the plugin had the
unexpected effect of making setuptools backend tests fail because of the file
finder hook setuptools_scm provides which includes unexpected files (because
they are under source control) in the sdist. To solve this, I have modified
those tests to copy the package files to a temporary location before building
so that the build source is not under git.