fix: change the way entry_points are mocked in metadata tests
In later versions of Python (>=3.11) the internals of working with entry points
have changed enough to break the mock based approach previously being used to
test dynamic metadata plugins. This commit replaces that with a version using
real EntryPoint objects which load real functions in the
test_dynamic_metadata.py file, the only mocking is switching the real list of
entry points with the list of test versions.