feat: allow metadata plugins to provide multiple values
This commit modifies the interface for metadata plugins to return a dictionary
of metadata keys and values, rather than just a single value. This is useful
for plugins which may run costly functions (e.g. CMake) that can provide multiple
metadata values, which will now only require a single plugin implementation and
a single run during the build.