feat: support dynamic metadata (#197)
This is a first pass at a scheme for supporting dynamic metadata in a
generic way. It keeps everything as simple as possible by using
functions in modules with a very simple signature to provide each value.
I am happy to discuss the proposed interface and make changes, but
thought it would be best to get an implementation out there to have
something concrete to discuss.

Scikit-build-core's settings have been extended to have an additional
metadata field which is just an optional list mapping string keys to
string values. For each entry in this list, if the key is also included
in the project.dynamic list, scikit-build-core will look up an a module
with a function. This should be a function that accepts the parsed
pyproject.toml (this seemed like a minimum but the final chosen
signature is very much up for debate, for example, maybe the config
settings dict should also be passed) and returns the dict that should be
merged into the pyproject["project"] entry in the loaded toml config.

Also included are a couple of core "plugins" which provide entry-points
for setuptools_scm and hatch-fancy-pypi-readme, essentially just
adapters as neither tool provides an explicit function matching the
proposed entry-point signature.

For tests, I have created a fixture which mocks the entrypoints list
with fake functions to give more direct control over what values we want
to test. I have duplicated most of the "simplest_c" test package to
allow the complete build process to be applied, but it is also possible
to test with only a pyproject.toml file, and in fact most of the tests I
have currently implemented do only use the pyproject.toml file and stop
at testing the calculated metadata, although this can of course be
changed. There is also a simple test for the setuptools_scm and
hatch-fancy-pypi-readme plugins to make sure they are working correctly.

Adding setuptools_scm as a test dependency to test the plugin had the
unexpected effect of making setuptools backend tests fail because of the
file finder hook setuptools_scm provides which includes unexpected files
(because they are under source control) in the sdist. To solve this, I
have modified those tests to copy the package files to a temporary
location before building so that the build source is not under git.

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: Ben Rowland <ben.rowland@hallmarq.net>
Co-authored-by: Henry Schreiner <henryschreineriii@gmail.com>