feat: experimental support for editable installs (#212)
This is a very early implementation of editable mode. One mode is
supported: redirect, which pulls the auto-discovered or manually
specified Python package files from the source locations, and the binary
files from the installed wheel. This method respects includes and
ignores, but doesn't adjust automatically if new files are added.

A few planned changes:
- [x] `_editable.py` will be renamed and moved to resources.
- [x] configure return changes will be reverted (at least for now)
- [x] utils in wheel will be cleaned up a bit
- [x] dataclass will be removed from Finder (minimal import impact
important)
- [x] A few unrelated changes will be pulled out into separate PR, #213
- [x] Consider name (rebuild might be an option, so "static" not ideal)
- [x] Add rebuild (opt in)

For later:

- Include file extensions when converting modules
- Check datafile support
- Check coverage support on editable installs

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>