ci: Add example tests to Fedora packaging (#378)
So I have designed the tests such that you can simply add a new test by
adding a `main.fmf` file in the corresponding file. Hopefully the files
are easy to read that it is easy to expand upon.

`fmf` format is basically an extension to yaml format (see
`swig/main.fmf` for an example) where it inherits all values from the
parent path. You can use commands like `tmt tests show` to view how the
values it takes.

---------

Signed-off-by: Cristian Le <cristian.le@mpsd.mpg.de>