feat: Cache build environments by content hash
Write an archive consisting of the build environment provided by pip,
and then extract it at a path containing the SHA256 hash of its
contents.
This prevents unnecessary rebuilds triggered by CMake thinking that
packages provided by pip in the build environment have been changed,
when it's only the temporary path that changed.
Because the path contains the hash of its contents, this is perfectly
safe since a changed build environment will change the hash.

Signed-off-by: Tobias Markus <tobias@miglix.eu>