feat: regex and dynamic-metadata rewrite (#457)
Based on https://github.com/scikit-build/dynamic-metadata/pull/1 (with
some updates). Fixes
https://github.com/scikit-build/scikit-build-core/issues/435.

This doesn't start using `tool.dynamic-metadata` yet, to give us room to
change that. Plus we have to back-compat support
`tool.scikit-build.metadata`, so might as well keep that as the only way
to set this for now.

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>