fix: parse CMake version strings containing '-' (#508)
Uses a simple `str.split('-msvc')` to extract the version number.
`'-msvc'` in particular was chosen to allow for forwards-compatibility
with future builds of MSVC CMake. I did not check for older versions as
I do not expect they will be any less compatible and this is still an
improvement on the current situation of no compatibility at all.

In my opinion, splitting on `-` should be sufficient and provide
compatibility with any other builds of CMake with custom version
strings. However, I did not do this as I'm not familiar with all of the
possible builds of CMake out there and wanted to ensure functionality.

---
Update: Now parses using `str.split('-')` for extended compatibility.
Thanks @henryiii

---
Resolves: #500

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>