feat: add inplace mode for editable installs (#553)
This adds support for editable.mode=inplace, which does an inplace CMake
build, along with a simple editable mode .pth file. This has a lot of
caveats (just like CMake inplace builds), but works better with some
workflows and tools. It's certainly better than copying the output file
back into the source, which is what some projects are currently doing.
:)

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>