tests: try comparing uncompressed files (#610)
Testing this to see if it fixes #596, alternative to #609. The SDists
linked seem identical except for minor compression differences, and this
might depend on exact zlib versions, so if we can compare uncompressed
hashes, that's good enough. We'll need to clarify that reproducibility
is only guaranteed on the unzipped tars. (If this works, we should
probably check the headers too)

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>