feat: setting CMake variables though environment variables (#565)
This PR allows reading environment variables to define CMake variable.
This is especially useful when the set of valid values which a CMake
variable can assume is either not known a priori, or is not finite, and
so `overrides` is not really helpful.

Simple example:

```toml
[tool.scikit-build]
cmake.define.FOO = {env = "FOO"}
```

Bumps `typing_extensions` to being required for Python 3.9+.

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: Henry Schreiner <henryschreineriii@gmail.com>