fix:: CMAKE_ARGS that may have spaces inside quotes (#727)
The current logic for parsing `CMAKE_ARGS` assumes that it is safe to
split the string on spaces. However, this may not be the case for all
variables; some variables must be passed with spaces in them, but inside
quotes. The
[`CMAKE_<LANG>_FLAGS`](https://cmake.org/cmake/help/latest/variable/CMAKE_LANG_FLAGS.html)
are a prime example of this. Quoting from the docs:
> This value is a command-line string fragment. Therefore, multiple
options should be separated by spaces, and options with spaces should be
quoted.

To support this, the split needs to be done with a regular expression
instead of a simple string split.

I left a note in the test regarding the fact that this behavior is
sensitive to quotes. I don't know if we can or should do anything about
that, though.

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: Henry Schreiner <henryschreineriii@gmail.com>