fix: support windows free-threading (#759)
Fixing the logic for finding a library to link to. Python fills the
LIBRARY slot with .dll for some reason. CMake was fixing this by finding
the `.lib` anyway, but now we feed it the right library to begin with,
which fixes free-threading builds. Tests are currently just downstream
in scikit-build-sample-projects until there's an easier way to get a
free-threaded build in GHA, especially for Windows.

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>