fix(setuptools): only inject logic if `cmake_*` keywords present (#768)
This is a sort of follow up to #413 and #414. In sage-on-gentoo we
discovered that the presence of scikit-build-core breaks the building
sagemath (https://github.com/cschwan/sage-on-gentoo/issues/783).
scikit-build-core also happens to be an indirect dependency for building
sagemath's documentation.
On the basis that the presence of setuptolls_rust does not break
sagemath's build, I studied the difference with scikit-build-core. I
noticed that the elements `bundled in the single
"finalize_distribution_options" entrypoint in scikit is split between
two different setuptools entrypoints.
This PR reproduce that setup and does a bit of clean up. The bundling of
the old _prepare_extension_detection and _prepare_build_cmake_command
into finalize_distribution_options being redundant after attributing
them directly to separate entrypoints. I also renamed the functions to
better characterise them.

Using this PR, sagemath builds without problems and pyzmq, which uses
scikit-build-core also builds properly.

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>