tests: fix test cmake config (#847)
The function `configure_args ` that is used to prepare the arguments in
`tests/test_cmake_config.py` does not consider that there might be a
`CMAKE_GENERATOR` set in the environment (e.g. set to `Ninja`):


https://github.com/scikit-build/scikit-build-core/blob/901b1d31d9b7b47ee93aa2ef6f887c15e82e7f7a/tests/test_cmake_config.py#L21-L30

The `configure()` in `src/scikit_build_core/cmake.py` mutates
`single_config` just right before calling the `Run(...).live(...)`


https://github.com/scikit-build/scikit-build-core/blob/901b1d31d9b7b47ee93aa2ef6f887c15e82e7f7a/src/scikit_build_core/cmake.py#L210-L212

It is not ideal to have tests affected by non-controlled environmental
variables (e.g. `CMAKE_GENERATOR`).

This PR introduces parametrization of three tests that fail otherwise on
Windows (see #843), and monkey-patching of the environment variable
`CMAKE_GENERATOR`.

Likewise, `sysconfig.get_platform()` is monkey-patched and parametrized
with `win` and `linux`.

The module `cmake` needs to be reloaded every time, because of the
`single_config` member of `CMakes` initialized in the factory for the
`dataclass`:
https://github.com/scikit-build/scikit-build-core/blob/901b1d31d9b7b47ee93aa2ef6f887c15e82e7f7a/src/scikit_build_core/cmake.py#L84

---------

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>