fix: use correct settings key to remove parts of the version in regex metadata plugin (#887)
Hello together.

I've been trying to use the `remove` settings key for the `regex`
metadata plugin and noticed, that it has been non-functional for me. To
me, it looks like the plugin acquires the wrong value from the settings
dictionary (`result` instead of `remove`). This PR intends to change
this to what I believe is correct.

---------

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>