docs: generate config value reference (#1052)
This would make it easier to cross-reference. I was debating going for
an auto-generated one and integrating with sphinx directly or going with
the `nox -t gen` files, and I think I prefer the latter because it makes
it more obvious to review the generated text in all the places, wdyt?

I also want to unlock usage of sphinx roles in it, but the readme
generation gets in the way. My opinion is that if we have this, we could
either move the readme part to the generated page, or truncate it to
only display the first summary sentence similar to how `click` does it.

Depends-on: #1051
Closes #999

---------

Signed-off-by: Cristian Le <git@lecris.dev>