feat: add a setting to disallow hard-coding some setting keys in the `pyproject.toml` (#1078)
This complements the discussion in #1050 that we should not allow some
fields to be hard-coded. We should allow them in the context of
overrides, so here is a rather convoluted way of adding that check.

This will also make it possible to extend it further if we want more
specific checks for specific overrides matchers as well. E.g. we can add
a restriction that `build.requires` cannot have `@` in an `sdist`

---------

Signed-off-by: Cristian Le <git@lecris.dev>
Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>
Signed-off-by: Henry Schreiner <henryfs@princeton.edu>
Co-authored-by: Henry Schreiner <henryschreineriii@gmail.com>
Co-authored-by: Henry Schreiner <henryfs@princeton.edu>