tests: rework downloading the wheelhouse for isolation tests (#1199)
Hopefully adding output capture will help us debug the intermittent
failure on 3.12 • CMake 3.26 on windows-latest.

The problem is each worker is trying to download the same wheelhouse.
I've reworked how the wheelhouse is setup, including a way to run the
setup standalone so that something like Fedora could pre-download it
before running pytest, no longer requiring the network. `python
tests/utils/download_wheels.py` should do it if pytest is normally
configured. We could put these in a top-level folder if that's better.

---------

Signed-off-by: Henry Schreiner <henryfs@princeton.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>